requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
python-docx
requests
